fix today norm
