+++ src/Store/currentUser/currentUserSelectors.js
@@ -1,3 +1,4 @@
 export const selectUserData = state => state.currentUser.user;
 export const selectIsLoading = state => state.currentUser.isLoading;
 export const selectError = state => state.currentUser.error;
+export const selectDailyNorm = state => state.currentUser.user.dailyNorm;
