+++ src/Store/currentUser/currentUserThunk.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { getCurrentUser } from 'services/api';
+import { getCurrentUser, updateDailyNorma } from 'services/api';
 
 export const getCurrentUserThunk = createAsyncThunk(
   'user/getCurrent',
@@ -14,6 +14,25 @@ export const getCurrentUserThunk = createAsyncThunk(
       return res;
     } catch (error) {
       return thunkAPI.rejectWithValue(error.message);
+    }
+  }
+);
+export const changeDailyNormaThunk = createAsyncThunk(
+  'waterNorma/change',
+  async (newNorma, thunkAPI) => {
+    try {
+      const token = thunkAPI.getState().auth.token;
+      newNorma = parseInt(newNorma);
+      console.log(`newNorma:`, newNorma)
+      if (newNorma <= 0 || newNorma > 15000 || Number.isNaN(newNorma))
+        throw new Error(
+          'Water norma must be more then 0ml, less then 15000ml and be a number'
+        );
+      const updatedDaliyNorma = await updateDailyNorma({dailyNorm: newNorma}, token);
+     
+      return updatedDaliyNorma;
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
